# Alternative scheduler with proper graph traversal