- LMS graph scheduling  
	- `lookupInner`: a mapping from `Sym` to `Node` with metadata, wrapped as `RepNode`  
		- priority value  
		- ~~`scheHere`~~`reachableHot`: required in the outer scope, at least once  
			- will be scheduled there only if available  
			- verb: `reach` vs `reachInner`  
		- ~~`reachHard`~~`reachableHard`: required by a hard dependency path, at least once  
			- unnecessary, will prune further reachability if unavailable  
			- we only care about the soft deps of a node actually scheduled in the outer scope  
			- verb: `reach` vs `reachSoft`  
	- `ScheduleQueue`  
		- maintain a descending priority queue  
		- no repeated enqueue for performance -- metadata update is in-place  
		- `lookup` has the same scope as `inner`, also serving as scheduling boundary  
		- optimization: share the same `lookup` for blocks sharing the same `inner`  
	- testing  
		- pipeline passed  
		- `echo` whole program: ~170s => ~35s  
		- `echo` application only: ~15s => ~5s  
		- `libc_kl_posix` library only: 8947s => ~71s  

To use, mixin `GraphTraversal` in your choice of transformer or codegen.