@Kraks 2023-02-15
> Should the standard code generators use this version by default, or should it be opt-in?

I think it looks good to just use this version by default.