@supunab 2023-02-15
Flare works as expected as well (ran all tpch queries and verified the output)